Installation docs including libcerf on macOS
I noticed during testing the installation on macOS from source that libcerf was not present what I tried to run the python interface. After speaking with Joachim, I realised that I need to get the libcerf code. The python interface appears to work with the code included as shown above.